// interfaces/ShippingAddress.ts
// interfaces/ShippingAddress.ts

export interface ShippingAddress {
    shipping_first_name: string;
    shipping_last_name: string;
    shipping_street_address1: string;
    shipping_street_address2: string;
    shipping_city: string;
    shipping_state: string;
    shipping_zipcode: string;
    shipping_country: string;
    shipping_phone: string;
}